% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Run the main analysis}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  maxCores = 1,
  outputFolder,
  databaseId,
  createCohorts = TRUE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{A database schema containing health care data in the OMOP Commond
Data Model. Note that for SQL Server, botth the database and schema
should be specified, e.g. 'cdm_schema.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema where the exposure and outcome cohorts will be
created.}

\item{cohortTable}{The name of the table that will be created to store the exposure
and outcome cohorts.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{outputFolder}{Name of local folder to place intermediary results; make sure to use
forward slashes (/). Do not use a folder on a network drive since
this greatly impacts performance.}

\item{databaseId}{A string used to identify the database in the results.}

\item{createCohorts}{Should the cohorts be created? If \code{FALSE}, the cohorts are assumed to already
exist.}
}
\description{
Run the main analysis
}
