% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcCorrelation.R
\docType{package}
\name{NcCorrelation-package}
\alias{NcCorrelation}
\alias{NcCorrelation-package}
\title{NcCorrelation: Estimating correlation between negative controls}
\description{
Evaluation of the performance of comparative effect estimation methods when sample size (number of subjects in the target and comparator) is small.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ohdsi-studies/NcCorrelation/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
