% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotCorrelations}
\alias{plotCorrelations}
\title{Plot correlation matrices}
\usage{
plotCorrelations(exportFolder, plotFolder)
}
\arguments{
\item{exportFolder}{The folder containing the CSV files.}

\item{plotFolder}{The folder where the plots will be written.}
}
\value{
Does not return anything. Is called for the side-effect of having plot files written.
}
\description{
Plot correlation matrices
}
