% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{writeExtremeCorrelations}
\alias{writeExtremeCorrelations}
\title{Write most extreme correlations to file}
\usage{
writeExtremeCorrelations(exportFolder, plotFolder)
}
\arguments{
\item{exportFolder}{The folder containing the CSV files.}

\item{plotFolder}{The folder where the table will be written.}
}
\value{
Does not return anything. Is called for the side-effect of having a CSV file written.
}
\description{
Write most extreme correlations to file
}
