% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Run the main analysis}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  maxCores = 1,
  minCellCount = 5,
  outputFolder,
  databaseId,
  createCohorts = TRUE,
  runCohortMethod = TRUE,
  doBootstrap = TRUE,
  computeCorrelation = TRUE,
  exportForSharing = TRUE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{A database schema containing health care data in the OMOP Common
Data Model. Note that for SQL Server, both the database and schema
should be specified, e.g. 'cdm_schema.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema where the exposure and outcome cohorts will be
created.}

\item{cohortTable}{The name of the table that will be created to store the exposure
and outcome cohorts.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{minCellCount}{Minimum cell count to report. Counts below this will be masked to protect privacy.}

\item{outputFolder}{Name of local folder to place intermediary results; make sure to use
forward slashes (/). Do not use a folder on a network drive since
this greatly impacts performance.}

\item{databaseId}{A string used to identify the database in the results. Will be used
to generate file names, so avoid special characters.}

\item{createCohorts}{Should the cohorts be created? If \code{FALSE}, the cohorts are assumed to already
exist.}

\item{runCohortMethod}{Run CohortMethod to produce effect-size estimates?}

\item{doBootstrap}{Perform the bootstrap? Requires CohortMethod to have completed.}

\item{computeCorrelation}{Compute correlation between negative control estimates? Requires
the bootstrap to have completed.}

\item{exportForSharing}{Export results to CSV files for sharing?}
}
\description{
Run the main analysis
}
