% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcCorrelation.R
\docType{package}
\name{NcCorrelation-package}
\alias{NcCorrelation}
\alias{NcCorrelation-package}
\title{NcCorrelation: Estimating Correlation Between Negative Controls}
\description{
An exploration into the correlation between negative control estimates.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ohdsi-studies/NcCorrelation/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
